Java entry point / main function
Restructure finding of the entry point of the Java function or the main function
of the program.

+++ src/java_bytecode/java_bytecode_language.h
@@ -69,10 +69,15 @@ public:
 protected:
   irep_idt main_class;
   java_class_loadert java_class_loader;
-  bool assume_inputs_non_null;
-  bool disable_runtime_checks;
-  int max_nondet_array_length;
-  int max_user_array_length;
+  bool assume_inputs_non_null;      // assume inputs variables to be non-null
+
+  bool disable_runtime_checks;      // disable run-time checks for java, i.e.,
+                                    // ASSERTS for
+                                    //  - checkcast / instanceof
+                                    //  - array bounds check
+                                    //  - array size for newarray
+  size_t max_nondet_array_length;   // maximal length for non-det array creation
+  size_t max_user_array_length;     // max size for user supplied arrays
 };
 
 languaget *new_java_bytecode_language();
